MB-33167 [SR]: DCP Producer sends DCP_ABORT
With this patch a queue_op::abort_sync_write item is streamed to a DCP
Consumer as a DCP_ABORT message.
Note that Abort items (as Prepare) are ignored for pre-MH replica
connections.

Note: Currently we send the Key in the DPC_ABORT payload. That is just
temporary and will be removed as soon as we have implemented the
DurabilityMonitor at Replica.

Change-Id: I265093bf130a633f2a58bf59f8da6c4eb8fe8d2b
Reviewed-on: http://review.couchbase.org/107089
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>