Refactor [SR]: Use Item::getCommitted only when strictly necessary
The main motivation behind this change is that the Item::getCommitted
function requires (*this) being an operation that can be translated to a
StoredValue's CommittedState.

But, a queue_op::abort_sync_write doesn't translate to any
CommittedState.
Given that Item::getCommitted is (indirectly) called by
ActiveStream::makeResponseFromItem when we build a DcpResponse to
stream, we need to avoid that call in that code path for making DCP_ABORT
messages.

With this patch I'm actually limiting the use of Item::getCommitted to
when it is strictly necessary, i.e. to when we set the CommittedState for
StoredValue from Item::op. That is to avoid any further similar issue.

Change-Id: I7dcf8bda408565bc2f19502135ab9ae4302bcc1a
Reviewed-on: http://review.couchbase.org/106880
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>