OpenTracing: Allow the clients to use Text encoding
The C++ API allows for a binary and a textual method to decode
the context received from another party. Unfortunately the decode
method don't detect the difference and force the user to know which
one to use.

It turns out that Jaeger don't implement the binary encoding in
all languages, and to work around that for our prototype this patch
adds support for trying to decode by using a text map in addition
to the binary encoding.

If the context contains '&' we assume that it is encoded as a text
map and try to decode by using a textual decoding. If that fails
it'll fall back and try to decode it by using the binary method.
If the context don't contain '&' we'll start off by using the
binary decode and fall back to a textual decode if that fails.

Change-Id: Icd3a6582f3122dec7766f1eb2a9b550501a9eb71
Reviewed-on: http://review.couchbase.org/106882
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>