MB-33333: Modify SeqnoAcknowledged to take prepared_seqno
Update DCP_SEQNO_ACKNOWLEDGED message and related methods to take a
single prepared_seqno argument instead of the previous in-memory and
on-disk seqnos.

Given the DurabilityMonitor still tracks both in-memory and on-disk
(and doesn't yet have a unified prepared_seqno), temporarily use the
'on_disk_seqno' for prepared_seqno. This is "correct", in that
anything which has been written to disk must have met its (local)
durability requirements, but is a pessimisation for level=majority or
majorityAndPersistActive. However it allows the interface to be
migrated to the new, single prepared_seqno and have the
DurabilityMonitor be updated in parallel / later.

A couple of tests had to be disabled while in this transition phase,
these will be re-enabled once DurabilityMonitor conversion is
complete.

Change-Id: I1150f634e6b3ec7c447eee799700228734ec0d63
Reviewed-on: http://review.couchbase.org/107082
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>