MB-33463: Implement getVBucketState and snapshotVBucket for magma
The initial implementation of magma only used the cached version of
vbstate. This checkin actually persists vbstate to the local DB.

Magma requires an additional data structure MagmaInfo to track things
that are normally tracked in couchstore. MagmaInfo is stored with
vbstate data in the localDB but in separate caches.

Change-Id: Ib1f14e86bbf2e67b6590bfaa4af19a8033d14788
Reviewed-on: http://review.couchbase.org/107141
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>