MB-33599: Allow dynamic reconfig of interfaces
Add support for adding / deleting or changing network
interface descriptions. To simplify the implementation
all reconfiguration happens in the dispatcher thread
and not as part of reloading the config file.

To avoid ending up in a situation where we no longer
can connect to memcached changing the interfaces
works by first enabling all new interface descriptions,
and if no error happens it'll shut down the ones
no longer in the configuration.

Change-Id: Id036b22bcfdc9780f1a653d9099d9750b354bdd6
Reviewed-on: http://review.couchbase.org/107291
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>