Fix unit test failure after moving to openssl 1.1.1b
memcached tries to disconnect the client with:

SSL_accept() returned -1 with error 1: error:1417C0C7:SSL routines:tls_process_client_certificate:peer did not return a certificate

But for some reason we don't detect an error in BIO_do_handshake like
we did before. We do however detect the problem the next time we
try to send a command.

Change-Id: I0ee5698fbe02b410fcdced86387a6bfdeafeacb2
Reviewed-on: http://review.couchbase.org/107366
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>