MB-33650 [SR]: Test bytes added in BufferLog for Durability DCP messages (take 2)
[Re-apply after isues with initial version - see MB-34350]

The base size for a DCP_COMMIT/DCP_ABORT message is 40 (i.e., 24
(header) + 16 (payload, which is the prepare + Commit/Abort seqno),
plus the size of the key.

The base size for a DCP_PREPARE message is 57, plus the sizes of key and
value. The base size (57) is similar to the base size of a DCP_MUTATION
(55), + 1 for the delte-flag, + 3 for Durability Requirements, - 2 for
missing optional-extras length.

This change:
- Adds checks on what we add as bytes-outstanding in the
  Producer::BufferLog when we send DCP_COMMIT/DCP_ABORT
- Adds the same check for DCP_PREPARE
- Removes unused code

Change-Id: Id5539e427aaba81e293a9e65c0ea1844c0f2348f
Reviewed-on: http://review.couchbase.org/111351
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>