MB-33650 [SR]: Test bytes added in BufferLog for Durability DCP messages
The base size for a DCP_COMMIT/DCP_ABORT message is 32 (i.e., 24
(header) + 8 (payload, which is the Commit/Abort seqno), plus the size
of the key.

The base size for a DCP_PREPARE message is 57, plus the sizes of key and
value. The base size (57) is similar to the base size of a DCP_MUTATION
(55), + 1 for the delte-flag, + 3 for Durability Requirements, - 2 for
missing optional-extras length.

This change:
- Adds checks on what we add as bytes-outstanding in the
  Prodicer::BufferLog when we send DCP_COMMIT/DCP_ABORT
- Adds the same check for DCP_PREPARE
- Removes unused code

Change-Id: Ia21c1e4f976c2dc02e6abd8793036017d7825b4a
Reviewed-on: http://review.couchbase.org/107367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>