MB-33684: Convert VBucket::stateLock to folly::SharedMutex
The initial motivation for this was change was part of debugging a
TSan issue, where it wasn't clear if ThreadSaniziter was correctly
tracking data-race issues between VBucket::stateLock (implemented
using pthread_rw_lock) and folly::SharedMutex.

The answer to that question was "yes"[1], however folly::SharedMutex
is a more performant mutex and the standard one used by
Synchronized<>, so generally better to use over the previous RWLock.

[1] Assuming you're not on macOS where the annotations TSan requires
    on mutexes aren't present...

Change-Id: If198f8a021066996c1206685558941c4aff2dd6c
Reviewed-on: http://review.couchbase.org/107374
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>