MB-33684: Convert VB::Manifest lock to folly::SharedMutex
The initial motivation for this was change was part of debugging a
TSan issue, where it wasn't clear if ThreadSaniziter was correctly
tracking data-race issues between VB::Manifest lock (implemented using
pthread_rw_lock) and folly::SharedMutex.

The answer to that question was "yes", however folly::SharedMutex is a
more performant mutex and the standard one used by Synchronized<>, so
generally better to use over the previous RWLock. It also permits
mutex upgrade which can be used to reduce the instances where
WriteHolder of the Manifest is required.

Change-Id: I787c15e2cd93962f7eaaf4560081d0e4bc88cf3a
Reviewed-on: http://review.couchbase.org/107375
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>