Modify vbucket_state_t to be exclusive enum, add RequestedVBState
Remove the `alive` member from the enumeration. This allows the enum to
be fully covered when used in most switch statements (e.g. switching on
vb->getState()), which currenlty isn't possible without having a dummy
entry for `alive` with the current enum definition.

(Note that vbucket_state_alive is not a valid state for a vBucket,
it's simply a mask used for some client requests to indicate they are
requesting any state apart from dead).

Move the use of Alive (for GET_ALL_VB_SEQNOS) into it's own explicit
type - RequestedVBState.

Change-Id: Idd1119d6f5ac35ca42c391264df945d58aebc5b2
Reviewed-on: http://review.couchbase.org/107380
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>