Remove recursive usage of VB::Manifest::mutex
EphemeralBucketStatTest.VBSeqlistStats recursively locks the
VB::Manifest lock as it creates both a ReadHandle and
CachingReadHandle. This is permitted for pthread_rwlock_t (as long as
it is acquired in read mode), but not for folly::SharedMutex (which we
want to migrate to).

Restructure the test to avoid the use of both Handles, by using the
same mechanism as the real PagerVisitor.

EventuallyPersistentEngine::getAllVBucketSequenceNumbers() has a
similar problem - it (recursively) locked the Manifest twice. This
case just looks like a simple coding error, just remove the extra
lock.

Change-Id: I09576f16b49b67d6bb27945e7ae733447d781873
Reviewed-on: http://review.couchbase.org/107471
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>