MB-33702: Don't create new failover entry if state still active
In KVBucket::setVBucketState(), if the old and new state are both
active then don't create a new failover table entry.

This scenario only recently became possible, as ns_server may send a
SET_VBUCKET_STATE(active) message when the vBucket is already active,
if it needs to update the replication topology. The topology should be
updated, but this shouldn't be considered a new branch in the vBucket
hisotry.

Change-Id: If9524f4e3393e5898927209efb50e76f5f814828
Reviewed-on: http://review.couchbase.org/107511
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>