MB-33389 [SR]: Move replicationTopology to VBucket
To be able to persist the replicationTopology it needs to be read by
the flusher when forming the new vBucket state. This also needs
correct locking to avoid data races. It therefore has a similar
ownership model as VBucket::state, which is owned by VBucket.

To handle synchronization, use folly::Synchronized<> to allow read and
write access, moving the topology to the VBucket as the central source
of truth (the DurabilityMonitor only needs specific data out of it).

Change-Id: I9ddf18698344660367a058b25a8e9e4ca8eeb66a
Reviewed-on: http://review.couchbase.org/107535
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>