MB-33372: Address expel case item after mutation with same seqno
An example is as follows:

     1000 - dummy item
     1001 - checkpoint start
     1001 - set VB state
     1001 - mutation
     1001 - checkpoint end  <<<<<<< dcpCursor1

Before the fix in this patch we would identify that we were at an item
that was equal to the highest seqno for the checkpoint and a we would
move backwards one place.  As we are now on a mutation we would not move
back further.  So we would expel the mutation (and previous items.)

Given that the expel pointer refers to the mutation with the highest
seqno in the checkpoint it should not expel it.  Hence in the above
example it should not expel any items.

Change-Id: Ia7d1a7abc63c447cbe2edcd6adfcd6f642876082
Reviewed-on: http://review.couchbase.org/107546
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>