[BP] MB-33906: Return NOT_MY_VBUCKET for getReplica on pending vbucket
[Backport of MB-33683.]

Currently when a GET_REPLICA operation is sent to a vbucket in a pending
state it is placed in the pendingOps list and is processed when the
vbucket is moved to the active state.

However given that is is not valid for a GET_REPLICA to be applied to
an active vbucket, when it gets to execute it will just NOT_MY_VBUCKET.

Another issue with placing a GET_REPLICA operation into the pendingOps
list is that if a rebalance fails and the vbucket move from a pending
state back into a replica state, the operation will be held in the
pendingOps list potentially indefinately.

This patch ensures that if a GET_REPLICA operation is sent to a vbucket
in a pending state we immediately return NOT_MY_VBUCKET.

Change-Id: Ifeab25a06ddb823c3e7285733d317d2623c4876e
Reviewed-on: http://review.couchbase.org/108159
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>