[SR]: De-duplicate SyncWrite opcodes correctly
The couchstore backend can store mutltiple documents for the same id
(as long as they are written in the same commit batch) - i.e. it
operates like a multimap and not just a map. (Note this isn't the case
for subsequent commit batches - a key modified there will replace any
previous instances of that key in prior batches).

This means that the ep-engine flusher *must* identify and de-duplicate
any items which should not be flushed to disk. This isn't just a
performance optimzation as is suggested by some of the code - failure
to de-duplicate will result in extra items being stored.

The current de-duplication algorithm in EPBucket::flushVBucket() works
by sorting all items first by key and them by seqno (descending), then
iterating across all items, comparing each with the "previous" one. If
they both have the same key then the current one (actually the one
with the lower seqno) is discarded as it is considered to have been
superceded.

However this is inadequate, because in general it assumes all items
with the same key can be de-duplicated, keeping only the one with the
highest seqno - for example:

    1:SET(A), 2:SET(B), 3:PRE(A)

Will only keep

    2:SET(B), 3:PRE(A)

As it considers seqno 1 to be superceeded (and can be de-duplicated)
by seqno 3. This however is incorrect, as we must keep the existing
(committed) value for a key when a prepared item occurs.

To address this we modify the sorting predicate and canDeDuplicate
function to also consider the namespace (i.e. if committed or
pending). Items are only de-duplicated if both the key and namespace
match.

Also expand a number of tests to check for these cases.

Change-Id: I021a4f184398be4f4017536b6cae3b3868555f9a
Reviewed-on: http://review.couchbase.org/107793
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>