Remove use of -1 as seqno in kvstore_test module tests
The TestDataStoredInTheRightVBucket test and
DelVBucketConcurrentOperationsTest test both use -1 as a seqno when
creating new items.  Currently we interchange between int64_t and
uint64_t when referring to a seqno, (for example, seqno in an item is
stored as a int64_t, however getHighSeqno / getLowSeqno in checkpoint
returns a uint64_t).  Overtime it would be good to use only an
unsigned int, as this is more consistent and it makes more sense for the
seqno to be a positive number. Therefore we want to avoid using -1 in
these tests.

In addition we do not want to repeatedly use the same seqno for a
sequence of non meta-data items as this does not represent the behaviour
of how seqnos are used.  Furthermore Magma can tolerate one batch with
all the same seqno's but if there are two batches with the same seqno
for all items, this will fail.  Therefore we want to increase the seqno
for each non-meta data item created in the tests.

Change-Id: Idcc47c983a4927cae3ff0ae65471b0ec5c03026f
Reviewed-on: http://review.couchbase.org/107777
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>