MB-33067: Fix intermittent failure in test_get_delete_missing_file
The test checks the behaviour of ep-engine when the database file is
missing by deleting the entire data directory. However this could
conflict with the ep-engine stats persistence task, resulting in the
test failing as it couldn't delete a file it previously listed:

    Running [0035/0047]: get/delete with missing db file
    (couchstore)...  Uncaught std::exception. what():cb::io::rmrf:
    remove of file
    ./ep_testsuite_basic.full_eviction.comp_passive.db\stats.json.new
    under ./ep_testsuite_basic.full_eviction.comp_passive.db failed:
    The system cannot find the file specified.

Modify the method used to make the database inaccessible - instead
just chmod() the file to be inaccessible without removng it.

Change-Id: I5de178c47b762778820c172a8f40940180fc8faf
Reviewed-on: http://review.couchbase.org/107809
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>