MB-33715 [SR]: Track only high_prepared_seqno in ADM
The original design for Durability allowed out-of-order commits.
To achieve that, the DurabilityMonitor at Replica needed to ack both
memory/disk seqnos, and the DM at Active had to track both as well.

After design review (see DesignDoc for motivation) we have to ensure
In-Order Commits. Now a Replica acks a single strictly-monotonic
quantity (the high_prepared_seqno), so the Active needs to track a
single seqno per Replica.

Note: we don't have a full implementation of high_prepared_seqno yet. So
(as for Replica) the Active acks a Prepare only when it is persisted
(even if it is Level=Majority). That is just a temporary (but
semantically correct) pessimization.

Change-Id: I41a59d3a314389c9bd77e139e57d306ba51e67f1
Reviewed-on: http://review.couchbase.org/107667
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>