MB-33637: Fix memcached testapp ChangeVerbosity LoggingTest
This test is currently flawed for multiple reasons.

1) When run in non-embedded mode (every CV run) the test looks will
always pass because the cb::logger::checkLogLevels() call will look
in the registry belonging to the testapp process and not the
memcached process. This means that it will find no loggers to check
the verbosity of, and will always return true.

2) The test log levels are the wrong way around.

3) The test is racey. We might attempt to shut down the memcached
server before the thread that changes verbosity completes execution.
This thread will segfault when it logs if this is the case.

Fix this test by:

a) Creating an EWB command that returns SUCCESS if the log levels are
the correct value.

b) Fixing the log level mapping.

c) Runing the verbosity change in the memcached worker thread. It's
a quick operation that will only block if we are registering or
unregistered loggers (creating/destroying buckets/DCP connections).
Verbosity changing is also very infrequently used.


Change-Id: I1b9d36217e7108b44663b8f5bcf013ceae6a4050
Reviewed-on: http://review.couchbase.org/107289
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>