MB-33407 [SR]: Add warmed up items to DurabilityMonitor
In addition to adding in-flight prepared SyncWrites to the HashTable
during warmup, also add them to the DurabilityMonitor.

This slightly adjusts the previous approach - previously warmup would
add any Prepared SyncWrites directly to the HashTable, and then remove
them from the HT if subsequently Committed.

This patch instead maintains an 'outstandingPrepares' map which
prepared items are added to, and if subsequently committed they are
removed from. At the end of the scan, the remaining contents of
outstandingPrepares is added to the HashTable and DurabilityMonitor
via EPVBucket::restoreOutstandingPreparesFromWarmup(). That avoids the
need to try to remove things from the DurabilityMonitor after being
added.

Adds new / expands existing tests for both Active and Passive DM
variants.

Change-Id: Ibbb96a7bb9f2f906217d36667f73d28d7ac824ea
Reviewed-on: http://review.couchbase.org/108015
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>