MB-33773: Don't take a reference to a buffered item
Taking a reference items in the buffer.messages structure is not safe as
a closeStream will clear the messages, making the reference invalid.

To fix the problem no reference is taken, instead we use std::move
and whilst the queue is locked, move the DcpResponse* to the
task's scope, now clearing the queue can't delete the object. Other
changes are made to fit with this move, such as moving back on a
OOM failure.

Testing covers the original fault and exercises a few related code
paths to be try and assert that under these situations we don't end
up with a flowControl leak.

Change-Id: Icfe6fc42ac2be0d2a76b42fa4a533f615d065194
Reviewed-on: http://review.couchbase.org/107961
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>