MB-33860 [SR]: Implement High Prepared Seqno (HPS) logic in PassiveDM
The HPS represents the last locally-satisfied Prepare on a node.
As per In-Order Ack/Commit, the HPS is updated within the
"durability-fence" constraint. See DesignDoc for details.

Practically, an HPS update is triggered in response of one of the
following events:
1) A new Prepare is queued into the PassiveDM. The new Prepare may be
  locally-satisfied immediatelly.
2) The Flusher has persisted some Prepare. That may "move" the
  durability-fence onward and satisfy a number of pending Prepares.

In follow-up patches, the HPS will be the quantity that Replica nodes
ack back to the Active.

Change-Id: I95b31c1cb15f929d20fcc2c67e23debd9755e7f8
Reviewed-on: http://review.couchbase.org/108027
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>