MB-33846: Test deleted system events don't change numDeletedItems
There was a bug in collections purgers where we would attempt to
decrement numDeletedItems below 0 when the pre item was a deleted system
event and the post item was passed as a nullptr (i.e. removed from the
hash table).

This bug was fixed by http://review.couchbase.org/#/c/108274/.

Add a test to guard against regression.

Change-Id: Id68be668c40519bd666e6456180f2e4d488289a8
Reviewed-on: http://review.couchbase.org/108028
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>