MB-33860, MB-33845 [SR]: SeqnoAck by High Prepared Seqno
The HPS is the quantity ack'ed by the PassiveDM.
The PassiveDM triggers a SeqnoAck(HPS) to the Active every time the HPS
is updated, which may happen as consequence of 2 different events:

1) A new Prepare is added to tracking, as the new Prepare may be
  locally-satisfied as soon as queued into the PassiveDM

2) The Flusher persists a flush-batch that contains at some Prepare, as
  there could be a pending (i.e., non locally-satisfied) PersistToMajority
  Prepare (i.e., a durability-fence).

The logic in the PassiveDM minimizes the number of SeqnoAck messages by
triggering them only when the HPS actually moves.

Note that this patch implicitly fixes MB-33845 (Double DcpSeqnoAck for
SyncWrite).

Change-Id: I98d21b0b0b22130a10eabe86625b7ab684b373c8
Reviewed-on: http://review.couchbase.org/108140
Reviewed-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>