MB-33860 [SR]: Introduce the SeqnoAckCallback
This is a pre-requirement for the PassiveDurabilityMonitor to call back
to VBucket for sending a SeqnoAck message on the PassiveStream.
Note that new callback is not used yet, so there is no test in this
patch. The callback will be used and tested in follow-up patches where
the PassiveDM will actually trigger seqno-acks at high_prepared_seqno
updates.

Change-Id: I03b1030257e1a92d96cc173e9bd260a6df0b346b
Reviewed-on: http://review.couchbase.org/108149
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>