MB-33836, MB-32903 [SR]: Implement High Prepared Seqno in ActiveDM
The HPS represents the last locally-satisfied Prepare on a node.
As per In-Order Ack/Commit, the HPS is updated within the
"durability-fence" constraint. See DesignDoc for details.

Practically, an HPS update at Active is triggered in response of one of
the following events:
1) A new Prepare is queued into the ActiveDM. The new Prepare may be
  locally-satisfied immediatelly.
2) The Flusher has persisted some Prepare. That may "move" the
  durability-fence onward and satisfy a number of pending Prepares.

The HPS implementation at Active differs from the one at Replica on
two main points:
a) Level:MajorityAndPersistOnMaster does enforce a durability-fence at
  Active (while it does not at Replica)
b) The internal implementation of the HPS at Active is what we used to
  know as the Active (local) tracking in FirstChain. So (differently
  from the PassiveDM) we don't need a new tracked-quantity in ActiveDM.
  With this patch we just fix the "move-semantic" of the Active
  tracking in FirstChain to be consistent with the HPS rules.

Note that this patch fixes also MB-32903 (Level::Majority writes are
committed only at flush if no-replica), as now the HPS is updated (if
possible per durability-fence rules) also at ADM::addSyncWrite. So, if
the Active is the only node then a Level:Majority write may be committed
immediately.

Change-Id: I70165239d5f50c241f757733d4f270d08dcdacc6
Reviewed-on: http://review.couchbase.org/108016
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>