MB-32734 [SR]: Add syncDelete flag to KVStore metadata
Prepared SyncDeletes require slightly different handling to "normal"
deletes - they represent a _potential_ delete assuming the key is
Committed - but are not technically deleted themselves.

This distinction is most seen on disk; where we need to persist
Prepared SyncDeletes such that they can be subsequently loaded from
disk (e.g. during warmup), but they should _not_ be subject to things
like tombstone puring which "real" deletes are.

As such, Prepared SyncDeletes should be persisted to disk as valid,
not-yet-deleted documents (in the separate Prepare namespace). To
achieve this, but still be able to distinguish them from SyncWrites,
we need to add a new flag in each KVStore's metadata. When reading
Prepared SyncDeletes from disk, if this flag is set then the returned
Item should be marked as a (Prepared) delete.

Change-Id: I15fcddc2fcbbc61a7b934d667c5d94cd4c29b49a
Reviewed-on: http://review.couchbase.org/108275
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>