MB-33918: Fix cursor dropping trigger issue
There was an issue where we were doing an integer divide in the
calculation used to determine whether cursor dropping should be
triggered and also in the calculation used to determine the amount of
memory to recover.

This resulted in cursor dropping being triggered too aggressively and
thinking it need to recover more memory than memory it could.

Change-Id: I4571b51739a3e778fcf5463be361db5e0fe92da9
Reviewed-on: http://review.couchbase.org/108266
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>