MB-32734 [SR]: Restore (committed) SyncDeletes correctly into HT
Once a Prepared SyncDelete has been Committed, it should exist on-disk
as CommitViaPrepare but be marked as deleted. If it needs to be
restored into the HashTable it should also have it's CommitState
restored, however that isn't currently the case.

Fix by setting the commitState of the restored StoredValue from the
(backfilled) Item.

Expand the DurabilityWarmupTest cases to check for this.

Change-Id: I9d3b4a9e554541d30d778f2a7c03e6c21aad3a2e
Reviewed-on: http://review.couchbase.org/108277
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>