MB-33936 [SR]: Convert PDM to ADM at set-vbstate Replica->Active
A PassiveDM receives in-flight Prepares from the Active and tracks them
until completion (commit/abort).

At vbstate-change Replica->Active, the old PassiveDM must takeover and
start acting as an ActiveDM.

This patch adds supports for converting a PassiveDM to an ActiveDM. The
convertion ensures that the Prepares tracked in the (old) PDM are
retained after the transition has completed.
Also, the new Replication Topology must be applied to all the in-flight
Prepares in the new ActiveDM.

Change-Id: I3dab246a82cdf3b67dc407d54d843ccc19d6524e
Reviewed-on: http://review.couchbase.org/108340
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>