Use cb::unique_item_ptr in dcp_message_producers
The dcp_message_producers interface passes abstract `item` objects for
the implementing class (normally memcached Connection) to send out. It
is is the responsiblity of the implementing class to release these
items. Currently this has to be done manually - the interface just
uses a raw `item*` for Item objects.

In addition to being error-prone (however known leaks exist), this
increases the boilerplate needed in other implmeentations (e.g. mocks
used for testing) - they must ensure they call engine->release(item).

Specifically, a subsequent patch adds a GoogleMock implementation of
dcp_message_producers for more flexible / powerful testing of
expecations of the DCP interface. In this case a large amount of extra
boilerplate (adding WillByDefault custom handlers to all methods)
would be needed.

Given a bunch of changes are needed one way or another to support
GoogleMock, update the interface to use the cb::unique_item_ptr (a
unique_ptr with a custom deleter). This gives RAII-style resource
management for `item` objects, removing the need to manually release
them.

Change-Id: I9ba77c4cc695e1f1d2944fd3f506198c54bdcb2e
Reviewed-on: http://review.couchbase.org/108332
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>