MB-33739: Introduce ConnStore
This class manages the lifetime of ConnHandler objects used within
the ConnMap. It will be integrated into the ConnMap in a future patch.

The ConnStore is a new class that managed the lifetime of ConnHandler
objects in regards to cookies and vBuckets. We have encountered bugs in
the past where the VbToConnection map has contained a stale reference
to a connection that no longer exists. Whilst we have fixed the issues
we have found, the use of a weak_ptr in the VbToConnection map saves us
from most issues. Ideally we would like to change this to a reference
as we have seen some minor performance gains from doing so. Given the
complexity of the code, ConnStore attempts to ensure that we can't hit
these bugs accidentally by tightly coupling the CookieToConnection map
to the VbToConns map.

Change-Id: I1f081c29334924366246676bdfda7e88f9716ba8
Reviewed-on: http://review.couchbase.org/108352
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>