MB-32734 [SR]: During flush, treat prepared SyncDeletes as not deleted
Recall that Prepared SyncDeletes should not be treated as deletes on
disk - they shouldn't be tombstoned and shouldn't contribute to delete
statistics.

While this is now the case (inside KVStore subclasses a prepared
SyncDelete is not written as a delete), EPBucket::flushOneDelOrSet()
was failing to check if the item was prepared, and hence was
incorrectly calling the PersistenceCallback as a delete, not as a set.

The consequences of this were that the statistics for numTotalItems
etc could underflow, triggering a crash:

    terminating with uncaught exception of type std::underflow_error:
    ThrowExceptionUnderflowPolicy current:0 arg:1

Fix by checking for isPending() when selecting what type of operation
to flush. Includes a test - expanding DurabilityEPBucketTest to
perform a SyncWrite+commit followed by SyncDelete+commit which
triggers the underflow.

Change-Id: Idbf4a4d4b72762a0c9348852917c3ce42bbf695f
Reviewed-on: http://review.couchbase.org/108355
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>