MB-33952 [SR]: Allow HPS unchanged in PDM::notifyLocalPersistence
Currently we expect that the HPS always moves at persistence. But that
is wrong.

The reason is that HPS may be updated at ADM::addSyncWrite for a
Level:Majority or a Level:MajorityAndPersistOnMaster Prepare (if there
is no durability-fence preventing that). So, when the Prepare is
persisted, then it could be new-HPS=prev-HPS.

Change-Id: Icaece332ff4731e34936b6b1579b3f7ac410fcdf
Reviewed-on: http://review.couchbase.org/108358
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>