MB-33846/MB-32596: Correct stat tracking of commits for existing values
Currently stat tracking for the commit of a sync write on a
pre-existing value is incorrect, it uses the StoredValue of the
pending sync write instead of the StoredValue of the old item to
calculate the stat update for committed -> [removed] item.

Update the stat tracking to use the StoredValue of the old item
instead.

Change-Id: I0e8c90de5b4355d0c6be82a2fd65d44bf50f178e
Reviewed-on: http://review.couchbase.org/108389
Tested-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>