MB-33919: Generate delete time for newly deleted items
Rather than allowing the existing expiry time to become the delete-time,
which is flawed as per the MB, allow the  VBQueueItemCtx to determine
how the delete-time should be set.

DCP consumer and the with-meta interface should not generate a
delete-time, the incoming 'replicated' value should be accepted. However
a value of zero is not allowed, and will be ignored/regenerated
regardless of the VBQueueItemCtx setting.

Genuine deletes/expirys should generate a new delete-time in-line with
the issues highlighted in the MB.

Change-Id: I7ea6ed8575eaa510a02cceb37a86628a28405fb5
Reviewed-on: http://review.couchbase.org/108393
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>