MB-32591: Estimate memory recovered from expelling
We want to be able to trigger expelling using the same mechanism that
is currently used for cursor dropping.

For cursor dropping we calculate how much memory we want to recover.
We then iterate through each vbucket performing cursor dropping
and removeClosedUnrefCheckpoints (which actually frees the memory).
Once we have freed the required amount of memory we break out of the
vbucket loop.

Given we want to apply a similiar mechanism we need to calculate the
amount of memory recovered by expelling.  It is important to note that
the value returned is an optimistic estimate (similar to the value
returned by removeClosedUnrefCheckpoints), as it assumes that each
queued_item is not referenced by anyone else (e.g. a DCP stream) and
therefore its reference count will drop to zero on exiting the function
allowing the memory to be freed.

Change-Id: I997576374dbc59923f0d08f7ca93e172a1af074a
Reviewed-on: http://review.couchbase.org/108453
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>